requirements.txt
pytest
Appium-Python-Client
allure-pytest
selenium